Fix a bug in ljmm_mremap().
The bug is triggered when ljmm_mremap() tries to shrink the
a mmapped block. The problem is about mistakenly calculating
the starting address of the stretch of memory that is going
to be disposed.

The unit testing suite does not have a case to catch the
memory shrinking. I will add a testing case tomorrow.